refactor: extract shared next-page navigation logic into GreenLanes concern
Both GreenLanes wizard controllers (MovingRequirementsController and
ApplicableExemptionsController) contained an identical handle_next_page
method that parsed a next-page URL, merged controller-specific query
params, and appended a cache-busting timestamp. The only difference
between the two was which params formed the base query hash.

Extracts the shared URI-parsing/merging skeleton into
GreenLanes::Concerns::NextPageNavigation. Each controller now implements
the template methods next_page_base_query_params (required) and
next_page_extra_query_params (optional, defaults to {}) to supply its
own params without duplicating the structural logic.

No behaviour change. All 974 request/controller specs pass.
