+++ app/controllers/green_lanes/moving_requirements_controller.rb
@@ -1,6 +1,7 @@
 module GreenLanes
   class MovingRequirementsController < ApplicationController
     include GreenLanesHelper
+    include Concerns::NextPageNavigation
 
     before_action :disable_switch_service_banner,
                   :disable_search_form
@@ -48,24 +49,12 @@ module GreenLanes
       }
     end
 
-    def handle_next_page(next_page)
-      uri = URI(next_page)
-      path = uri.path
-      next_page_query = if uri.query
-                          Rack::Utils.parse_query(uri.query)
-                        else
-                          {}
-                        end
-
-      query = {
+    def next_page_base_query_params
+      {
         commodity_code: moving_requirements_params[:commodity_code],
         country_of_origin: moving_requirements_params[:country_of_origin],
         moving_date: @moving_requirements_form.moving_date.iso8601,
-      }.merge(next_page_query)
-       .merge(t: Time.zone.now.to_i)
-       .deep_symbolize_keys
-
-      "#{path}?#{query.to_query}"
+      }
     end
   end
 end
